Thoroughly check the reverted project to make sure working correctly  - Regression stats are gathered for each dataset seperately

+++ superhuge/models/commons/classify_interface.py
@@ -44,7 +44,7 @@ class ClassifyInterface(MInterface):
 
     def training_closure(self, data):
         outputs: torch.Tensor = self.forward(data)
-        targets = data["label"]
+        targets = data["label"].to(torch.long)
 
         return outputs, targets
